Following things are included in this commit.
1) Added the UI's for the configuration entities introduced in previous commit.
This will help to avoid the restarting of server. (Previously values were maintained in the property files) 

2) Renamed those two configuration entities to more closely follow the GoogleCheckout naming pattern.
The idea behind renaming of entity is that in future we can have other feature coming from Google so at that time GoogleConfiguration entity will be bit confusing.IMO this entities are introduced before few days(1 or 2 days) and I am sure no one is using these entities in Production so we will be safe by this change in the initial phase of google checkout.

Thanks to myself :)

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/specialpurpose@780501 13f79535-47bb-0310-9956-ffa450edef68