A patch from Anne Jessel "Improve handling and display of additional product images" (https://issues.apache.org/jira/browse/OFBIZ-4024) - OFBIZ-4024
This patch adds several closely-related features, intended to improve the display of products and their images in the ecommerce webapp.

When additional images for a product are uploaded via the Catalog, the scaled versions that were already being created in service addAdditionalViewForProduct are now recorded as ProductContent. This makes them easily available for use elsewhere.

New ProductContentType values were added to identify these content types. The obviously consistent string to use for these was too long, and so I had to choose something less consistent, but shorter and hopefully still clear. I used XTRA instead of ADDITIONAL.

The sizes to which additional images are rescaled has been changed, so they better fit in the ecommerce display. Relevant sizes in productdetail.ftl have been updated to match these sizes.

The ecommerce product display now uses the scaled additional images if available, rather than relying on the web browser to scale the images.

When displaying a single product in ecommerce, the main image is larger than the additional images. Hovering over an additional image displays a larger version in place of the main image.

Clicking on an additional image displays a popup with a detailed version of the image.

To use the new features, one or more large images (suggest at least 600x600) should be attached as additional images to a product, using the form on the bottom of the content tab for a product in Catalog. Viewing the product in ecommerce should then exhibit the above behaviour.

I have tried to maintain backwards compatibility, so that products with the old-style use of additional images will continue to display as they did previously. This means some of the code is not as streamlined as it might be.

JLR: You might prefer to keep the old ImageProperties.xml values...

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/specialpurpose@1042024 13f79535-47bb-0310-9956-ffa450edef68