BUG - PROJECTADMIN cannot create projects in projectmgr (https://issues.apache.org/jira/browse/OFBIZ-3201).
I add a new permission check to the ProjectMenu.xml to check weather the user have PROJECTMGR_ADMIN rights or not. If not he should not be able to create new Projects. 
That means that users which are in the security group PROJECTADMIN aren't allowed to create new projects by default. Because they are responsable for there own projects and not for creating new once. Maybe the name of the Permission group is a little bit confusing. The PROJECTADMIN is not the admin from the project module he is more a Project manager (an admin for his own projects).

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/specialpurpose@1087680 13f79535-47bb-0310-9956-ffa450edef68