A patch from Tom Burns for "eCommerce Create New Customer Improvement" https://issues.apache.org/jira/browse/OFBIZ-4831
1. Add support for dependent Country / State drop down to newcustomer.ftl
This change makes the behavior of the Country / State drop downs in newcustomer.ftl consistent with the Profile > Contact Information > Postal Address > "Create New" and "Update" behavior.

2. When an eCommerce user, who is not Logged In, clicks "Register" the newcustomer.ftl from opens with a "Go Back" button.
This button navigates to a Login form that prompts the user for a user name and password.
This makes no sense in the context.
A better design is to have a Cancel button. When the user clicks Cancel the application returns to the Main display for the current session.

jleroux: 
1. I wondered about returning to main page when coming from login screen and asking to create a new user. But it makes sense as well and is better when coming directly from main page. Anyway this is purely cosmetic
2. We don't need to get back to this screen (new profile) so no needs to use from-field="selectedStateName" in
<set field="selectedDependentOption" from-field="selectedStateName" default-value="_none_"/>
It makes things clearer

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/specialpurpose@1331842 13f79535-47bb-0310-9956-ffa450edef68