Fixes "Under some conditions the Paypal button encroachs the menu bar" https://issues.apache.org/jira/browse/OFBIZ-3136
Small improvement to provide a default in French when any country has been selected. The same idea can be used with other default locales.

Actually this issue can't be completely fixed without efforts to reduce the lengths of labels above buttons in some languages, also depends on screen widths and anyway OFBiz is far to provide a responsive design, if even this is wanted...

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/specialpurpose@1535981 13f79535-47bb-0310-9956-ffa450edef68