Fixes: jquery.jgrowl.min.css is linked twice in html page (OFBIZ-7758)
I am running on trunk rev 1751489 on tomahawk theme. 
In the browser, right-click and view the page source. 
There is duplicate linking of jquery.jgrowl.min.css i.e.

<link rel="stylesheet" href="/images/jquery/plugins/jquery-jgrowl/1.4.1/jquery.jgrowl.min.css" type="text/css"/>

The solution is to remove a hack for JPublish in fex Java classes 
and related changes (mostly "page." in ftl and xml files)

Thanks: Rohit Koushal for the path and James Yong the report

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/specialpurpose@1761235 13f79535-47bb-0310-9956-ffa450edef68