Fixed: After creating a new Flexible Birt report, before visualising it you need to do other actions (OFBIZ-9194)
The solution I provided with revision: 1781218 was wrong. A really satisfying
solution would be to gray the "Send" button on the 
birt/control/CreateFlexibleReport page. But that's not something easy to do
because it's generated from a content. 

So I decided to remove the part in the PrepareBirtCall.groovy and add an 
indication in the BirtErrorNotPublishedReport for users to use their browser 
back button. Because you catch a ViewHandlerException in 
RequestHandler.renderView() and there is not much you can do at this stage...

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-plugins/trunk@1788020 13f79535-47bb-0310-9956-ffa450edef68