Fixed: Implement and demonstrate few services in Birt Report Builder (OFBIZ-9192)
It was a long time I did not revisit this issue. I no longer stumbled upon 
issues I reported before in the Jira but found that 2 entities were missing.

I was an easy fix: the ServiceReportsView.xml was missing

Thanks: To Nicolas who in the meantime certainly committed fixes for the other 
issues I reported in the Jira

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-plugins/trunk@1789159 13f79535-47bb-0310-9956-ffa450edef68