Fixed: Implement and demonstrate few services in Birt Report Builder (OFBIZ-9192)
Jacopo reported an issue on dev ML on the Party view profile 
The InvoiceItemProductAndParty view use a partyName from the Party
entity, but there is no partyName in the Party.

I simply commented out this line in the view definition. 
PartyNameView could be used later there.

I also fixed partyName in BirtMasterReportServices class,
formatted it and used Debug.logError instead of printStackTrace()

Thanks: Jacopo for report

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-plugins/trunk@1790499 13f79535-47bb-0310-9956-ffa450edef68