Implemented: Remove dependencies from OFBiz framework to plugins (OFBIZ-9322)
This is related to a commit in ofbiz-framework r1792866 in which data 
that is removed in here is added to the framework. There is one modification
to the DemoPurchasing.xml file in which we added a UserLoginSecurityGroup which
is moved from DemoOrderPeopleData.xml because it has dependencies on security
data in ecommerce. Other than that, all data is simply deleted.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-plugins/trunk@1792867 13f79535-47bb-0310-9956-ffa450edef68