Fixed: Birt plugin depends on example plugin (OFBIZ-9383)
The BIRT demo data for flexible reports depends on the Example entity from the 
Example plugin.
We should either provide BIRT's own example data with example reports OR 
setup a gradle dependency on the example plugin so that everything is in place.

jleroux: I did not like to duplicate the example data in birt or an alike 
solution.
So I put a gradle dependency on the example plugin. 
If the example plugin misses the birt component can't build. 
I put some comments to explain the possible issue and its solution out of 
demonstration.

Thanks: Michael Brohl for report

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-plugins/trunk@1799699 13f79535-47bb-0310-9956-ffa450edef68