Fixed: Lucene tests fail due to problems in the test data. (OFBIZ-9439)
The Lucene test data references to the ownerContentId "CNTGIZMOS" which is not 
defined in the lucene demo data. It was instead defined in the ecommerce demo 
data in the file DemoWebsitePublishPointData.xml. The data is rewritten so that
it is usable for Lucene and does not interfere with other data.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-plugins/trunk@1800322 13f79535-47bb-0310-9956-ffa450edef68