Improved: Upgrade Jquery 1.11.0 to JQuery 3.2.1 (OFBIZ-9978)
Includes following changes
jQuery UI upgraded to 1.12.1
Included CSS files to theme properties and screens.

Manually added datepicker-en.js and datepicker-it-CH.js a new candidate added 
it to jsLanguageFilesMapping
Updated defaultDateTime path of dateTime picker it was having path of 
timepicker files.

I18n files for datepicker renamed from jquery.ui.datepicker- to datepicker- 
as comes with the JQuery ui package so that we don’t have to rename files 
manually everytime we upgrade.
https://github.com/jquery/jquery-ui/tree/master/ui/i18n
Can’t get jquery-ui-i18n.js file though it was not used either.

I have not updated jQuery.cookie-1.4.0.js as it is related to jstree and we will 
update & move it when we will update jstree.

Jeditable 1.7.2 upgraded to 1.7.3

I feel JQuery ui should also be in plugins folder with other JQuery plugin

Thanks: Aditya

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-plugins/trunk@1817026 13f79535-47bb-0310-9956-ffa450edef68