Fixed: Birt plugin depends on example plugin (OFBIZ-9383)
The BIRT demo data for flexible reports depends on the Example entity from the 
Example plugin.

We should either provide BIRT's own example data with example reports 
OR 
setup a gradle dependency on the example plugin so that everything is in place.

jleroux: this simply duplicates example data in birt component. There are
cases where duplication is the simplest solution even at the cost of possible
divergence in the future (not a problem for the birt component, so it's OK)


Thanks: Michael Brohl for report and Taher for discussion

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-plugins/trunk@1820973 13f79535-47bb-0310-9956-ffa450edef68